requirements.txt
aiogram
aiohttp
phonenumbers
python-dotenv
sqlalchemy 
aiosqlite

gspread 
google-auth-oauthlib 
pandas